apps/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "GM Smart Shield API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=0.4.0",
    "fastapi>=0.100.0",
    "httpx>=0.28.1",
    "langchain-text-splitters>=1.1.1",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-exporter-prometheus>=0.60b1",
    "opentelemetry-instrumentation-fastapi>=0.60b1",
    "opentelemetry-sdk>=1.39.1",
    "pandas>=3.0.1",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pypdf>=6.7.3",
    "python-dotenv>=1.0.0",
    "sentence-transformers>=5.2.3",
    "sqlalchemy>=2.0.0",
    "structlog>=25.5.0",
    "uvicorn>=0.20.0",
    "langchain>=0.3.0",
    "langchain-ollama>=0.2.0",
    "fastapi-mcp>=0.1.0",
    "ollama>=0.3.0",
    "PyYAML>=6.0.0",
]

[dependency-groups]
dev = [
    "httpx>=0.27.0",
    "lefthook>=1.0.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-bdd>=7.0.0",
    "ruff>=0.3.0",
]
